chore: Convert `api.planx.uk/send/bops.js` to TypeScript (#1434)
