+++ api.planx.uk/.eslintrc
@@ -38,7 +38,6 @@
     "clearTimeout": "readonly",
     "Buffer": "readonly",
     "Headers": "readonly",
-    "querystring": "readonly",
     "URLSearchParams": "readonly",
     "exports": "readonly",
     "fetch": "readonly",
