+++ api.planx.uk/helpers.ts
@@ -189,6 +189,7 @@ const makeUniqueFlow = (flowData: Flow["data"], replaceValue: string): Flow["dat
 };
 
 const isLiveEnv = () => (["production", "staging", "pizza"].includes(process.env.NODE_ENV || ""));
+
 /**
  * Get current environment, formatted for display
  */
